requirements.txt
streamlit==1.44.0
groq==0.15.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
